+++ invoke_plugin_for_pylint/_plugin.py
@@ -1,36 +1,28 @@
-# pylint:disable=too-many-arguments,too-many-boolean-expressions
+# pylint:disable=too-many-boolean-expressions
 """Pylint invoke plugin."""
 
 from __future__ import annotations
 
 from typing import Any
 
-import pylint
 from astroid.nodes import Arguments, AssignName, FunctionDef, NodeNG
 from pylint.checkers import BaseChecker
 from pylint.checkers.variables import VariablesChecker
-from pylint.interfaces import IAstroidChecker
 from pylint.lint import PyLinter
 
-_PL14 = pylint.__version__ > "2.14"
 
-if pylint.__version__ < "2.11":  # pragma: no cover
-    raise RuntimeError(f"pylint 2.11+ is required but found {pylint.__version__}")
-
-
-def _inject(checker: BaseChecker, linter: PyLinter) -> None:
+def _inject(linter: PyLinter) -> None:
     original = VariablesChecker.add_message
 
     def _new_add_message(
         self: Any,
         msgid: str,
+        *args: Any,
         line: Any = None,
         node: NodeNG | None = None,
-        args: Any = None,
-        confidence: Any = None,
-        col_offset: Any = None,
+        **kwargs: Any,
     ) -> None:
-        config = linter.config if _PL14 else checker.config
+        config = linter.config
         decorators = config.additional_invoke_task_decorators + [
             "invoke.tasks.task.inner",
             "invoke.tasks.task",
@@ -47,9 +39,9 @@ def _inject(checker: BaseChecker, linter: PyLinter) -> None:
             )
         ):
             return
-        original(self, msgid, line, node, args, confidence, col_offset)
+        original(self, msgid, line, node, *args, **kwargs)
 
-    VariablesChecker.add_message = _new_add_message
+    VariablesChecker.add_message = _new_add_message  # type:ignore[method-assign]
 
 
 class InvokeChecker(BaseChecker):  # type:ignore
@@ -57,9 +49,6 @@ class InvokeChecker(BaseChecker):  # type:ignore
 
     The checker will prevent 'unused-argument' for invoke task context arguments.
     """
-
-    if not _PL14:
-        __implements__ = IAstroidChecker
 
     name = "invoke-plugin-for-pylint"
     priority = -1
@@ -77,9 +66,9 @@ class InvokeChecker(BaseChecker):  # type:ignore
         ),
     )
 
-    def __init__(self, linter: PyLinter = None):
+    def __init__(self, linter: PyLinter):
         super().__init__(linter=linter)
-        _inject(self, linter)
+        _inject(linter)
 
 
 def register(linter: PyLinter) -> None: