+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# 3.0.0
+* Minimum required pylint version is 2.17
+* Support python 3.12
+* Add support for pylint 3.x
+* Dropped support for python 3.8
+
 # 2.0.0
 * Removed support for invoke 1.x